DriverUI: workaround chromium bug
chromium often sends the disconnect message too late which leads to the
player being considered leaving while still being in the driver UI. To
workaround this the player now only gets removed from the waiting queue,
if the disconnect message occurs.

In consequence leaving the DriverUI itself while controlling a vehicle
will *not* free the vehicle for other players!