WaitingQueue: change implementation to store the player in the vehicle
this rework is needed to allow other vehicle types since the old
implementation would store the BLE-UUID and wait until a player connects
until the BLE connection was established. This wasn't portable as it only
takes physical vehicles into account.

Furthermore with this rework a bug was fixed that the StaffUI would lose
it's UUID-Player map table when a driving site was loaded