fix controls not showing for assigned players
The old logic only tried getting a player vehicle, if a change was made in the queue.

However if a player already had a vehicle assigned, no change would oocur in the queue and the player wouldn't see the controls despite having an assigned vehicle and therefor permissions to control it.

This is now fixed