FullTrack: Rewrite building
The old version broke when curve pieces had another size or when adding straight pieces with other lengths. This was because the offset was calculated via the top left point of a "invisible square" around the piece.

Now the middle of the piece is taken into account which allows pieces with different sizes to be put together into a track